fix: fixing drop menu and create function to random colors
